[XTENSA] Misc Minor bug-fixes to get SMP kernel working correctly.
1. Initialization of traps has to be done after we call
   setup_per_cpu_areas(). Using a early exception table
   to handle the early hardware probes.

2. Added some debug code:

   a. Breakpoint functions for stack overflows in IRQ code.

   b. Code to check if preemption is disabled after doing
      a system call; use to be common in UNIX to find system
      calls that forgot to release a spinlock to set the interrupt
      level back to normal.

3. Fixed a bug in parsing boot params when not coming from U-Boot.

4. Fixed a couple lint errors in memmove.c; it's not called very
   often.

5. The three core HiFi-2 VARIANT has a problem where we only told
   ccbuild to pass 6 interrupts to the interrupt distributor. In
   fact the audio subsystem is using IRQ's {7, 9, 10, and 11}.
   Added a workaround timer to check the FIFO Level and calling
   the audio output interrupt handler.

6. With interrupts not being monotonically increasing, the previous
   approach of just adding three to the IRQ numbers for SMP didn't work;
   switched to using a separate set of assignments for SMP case.

7. Added a couple #defines to pgtable.h to allow some of the new
   PCI enabled code to compile.

Hopefully this kernel will survive the current overnight test.

Currently the this branch now runs the same kernel on the LX60, LX110,
and LX200 boards with V2 and V3 MMU's. Only significant problem
remaining is audio on the LX110 not being initialized 100% correct.

Symptoms are that the sound isn't coming out of the DAC and the
FIFO isn't applying back-pressure during the module initialization
when it plays a two second test sine wave. Will look into this later.

+++ arch/xtensa/kernel/pci.c
@@ -79,9 +79,10 @@ pcibios_align_resource(void *data, struct resource *res, resource_size_t size,
 		resource_size_t start = res->start;
 
 		if (size > 0x100) {
+			/* REMIND: 'dev->resource - res' make an int */
 			printk(KERN_ERR "PCI: I/O Region %s/%d too large"
 			       " (%ld bytes)\n", pci_name(dev),
-			       dev->resource - res, size);
+			       dev->resource - res, (long) size);
 		}
 
 		if (start & 0x300) {
